1. Type conversion of target label tensor for gold example. 2. Revised batch_size according to V100 GPU. 3. Set min_workers to zero. 4. Passing the device to dataset __init__ function. 5. Passing the missing parameters in validation loop for self.model.

+++ dataset.py
@@ -165,7 +165,7 @@ class VWSDDataset(Dataset):
         imgs = torch.cat([items[idx][1]['pixel_values'] for idx in range(batch_size)])\
             .to(self.device, non_blocking=self.non_blocking)
 
-        gold_examples = torch.Tensor([items[idx][2] for idx in range(batch_size)])\
+        gold_examples = torch.Tensor([items[idx][2] for idx in range(batch_size)]).type(torch.LongTensor)\
             .to(self.device, non_blocking=self.non_blocking)
 
         return text_inputs, imgs, gold_examples
